requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
asyncmy==0.2.9
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
google-generativeai>=0.8.0
weasyprint==60.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pydantic==2.5.0
pydantic-settings==2.1.0
pydyf<0.9.0
httpx>=0.27.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
beautifulsoup4==4.12.2
lxml==4.9.3

# RAG Chat Dependencies
chromadb>=0.5.0
sentence-transformers>=2.2.2
